review comments fixes
