+++ app/client/src/workers/common/JSLibrary/index.ts
@@ -34,12 +34,6 @@ export const defaultLibraries: TJSLibrary[] = [
     docsURL: "https://github.com/digitalbazaar/forge",
     name: "forge",
   },
-  {
-    accessor: ["deep-diff"],
-    version: "1.0.2",
-    docsURL: "https://github.com/flitbit/diff#readme",
-    name: "deep-diff",
-  },
 ];
 
 export const JSLibraries = [...defaultLibraries];
