+++ app/client/src/workers/common/JSLibrary/resetJSLibraries.ts
@@ -5,12 +5,10 @@ import forge from "node-forge";
 import { defaultLibraries } from "./index";
 import { JSLibraries, libraryReservedIdentifiers } from "./index";
 import { invalidEntityIdentifiers } from "../DependencyMap/utils";
-import deepDiff from "deep-diff";
 const defaultLibImplementations = {
   lodash: _,
   moment: moment,
   xmlParser: parser,
-  "deep-diff": deepDiff,
   // We are removing some functionalities of node-forge because they wont
   // work in the worker thread
   forge: /*#__PURE*/ _.omit(forge, ["tls", "http", "xhr", "socket", "task"]),
